local/requirements.txt
PyYAML
jmespath
pyauto.core>=0.0.6
pyauto.util>=0.0.3a3
